2/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
requests>=2.26.0
python-dotenv>=0.19.0
pydantic>=1.8.2
aiohttp>=3.8.1
openai>=1.0.0  # OpenAI API 客户端 